requirements.txt
# requirements.txt
flask
flask-cors
ultralytics
Pillow
gunicorn